use float32 for labels

+++ src/data.py
@@ -112,7 +112,7 @@ def build_preprocessor() -> ColumnTransformer:
 class DiabetesDataset(Dataset):
     def __init__(self, X, y):
         self.X = torch.tensor(X, dtype=torch.float32)
-        self.y = torch.tensor(y, dtype=torch.long)
+        self.y = torch.tensor(y, dtype=torch.float32)
 
     def __len__(self):
         return len(self.y)
